Create auraDataUpdating
this is a further updated version of the auradata json (staticData.json) from the dolphSol Macro made by BuilderDolphin. I made this for personal use, and did not write a majority of the code therefore I take little to no credit for the works of this file and credit goes to BuilderDolphin themselves.